swagger setup

+++ src/users/users.service.ts
@@ -1,34 +1,54 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto.js';
 import { UpdateUserDto } from './dto/update-user.dto.js';
 import { UserRepository } from './users.repository.js';
 import { User } from './entities/user.entity.js';
+import { UserResponseDto } from './dto/user-response.dto.js';
 
 @Injectable()
 export class UsersService {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async create(createUserDto: CreateUserDto) {
+  async create(createUserDto: CreateUserDto): Promise<User> {
     const createUser = {
       ...createUserDto
     } as User;
   
-    return this.userRepository.create(createUser);
+    const user = await this.userRepository.create(createUser);
+  
+    return user;
   }
 
-  async findAll() {
+  async findAll(): Promise<User[]> {
     return this.userRepository.findAll();
   }
 
-  async findOne(id: string) {
-    return this.userRepository.findOne(id);
+  async findOne(id: string): Promise<User | undefined> {
+    const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    return user;
   }
 
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  async update(id: string, updateUserDto: UpdateUserDto): Promise<User | undefined> {
+    const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
     return this.userRepository.update(id, updateUserDto);
   }
 
-  async remove(id: string) {
+  async remove(id: string): Promise<boolean> {
+    const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
     return this.userRepository.remove(id);
   }
 }
